HADOOP-12731. Remove useless boxing/unboxing code. Contributed by Kousuke Saruta.
(cherry picked from commit 736eb17a796a1c1ad5f4db2c6a64f6752db7bec3)
(cherry picked from commit 4edd7bd82094dd8c1474a912532e6373cedf79c9)

+++ hadoop-common-project/hadoop-auth/src/test/java/org/apache/hadoop/security/authentication/util/TestZKSignerSecretProvider.java
@@ -35,7 +35,7 @@ public class TestZKSignerSecretProvider {
 
   // rollover every 2 sec
   private final int timeout = 4000;
-  private final long rolloverFrequency = Long.valueOf(timeout / 2);
+  private final long rolloverFrequency = timeout / 2;
 
   @Before
   public void setup() throws Exception {
